firmware: migrate smc command to new PSCI firmware driver

+++ arch/arm/include/asm/psci.h
@@ -106,7 +106,7 @@ static inline void psci_set_ops(struct psci_ops *ops)
 }
 #endif
 
-long psci_invoke(ulong function, ulong arg0, ulong arg1, ulong arg2);
+u32 psci_invoke(ulong function, ulong arg0, ulong arg1, ulong arg2);
 
 void psci_cpu_entry(void);
